Introduce preemption-based, idleless service guarantees
If several processes are competing for the device at the
same time, but all processes and groups have the same weight,
then the mechanism introduced by this commit enables BFQ to
guarantee the expected throughput distribution without ever
idling the device. Throughput is then much higher in this
common scenario.

This commit also refactors the long function bfq_add_request,
and improves and extends most comments.

Change-Id: I0688647b593851695cfad6e24894550cf88e2efe
Signed-off-by: ben863 <r.budhiono@gmail.com>