Fix bug in optimized update of next_in_service entity
There was a case where bfq_update_next_in_service did not update
next_in_service, even if it might need to be changed: in case of
requeueing or repositioning of the entity that happened to be
pointed exactly by next_in_service. This could result in violation
of service guarantees, because, after a change of timestamps for
such an entity, it might be the case that next_in_service had to
point to a different entity. This commit fixes this bug.

Signed-off-by: Paolo Valente <paolo.valente@linaro.org>
Change-Id: I9196b424aad5a05962e658e815d8c1c8793dc112
Signed-off-by: ben863 <r.budhiono@gmail.com>